docs: 📝 Add rule overwriting information in `fields.md`
This commit adds a warning in the `fields.md` document of the API that describes how subsequent calls to the `rules` or `required` methods concatenate rather than overwrite rules. It then explains how to overwrite or reset them using `overwriteRules` method. This important information was added to help developers understand this behavior and avoid potential confusions or algorithm errors in the future.