feat: Relations filtering
The changes in this commit firstly introduce the functionality of adding matchable attributes to the EagerField in LaravelRestify.
In Filters/MatchesCollection.php, the collection method from the request input has been modified to append on all keys and their inputs.
In Traits/InteractWithSearch.php, matchable fields from the repository are now added to the MatchesCollection.
The functions onlyMatchable() and matchable() have been added to RelatedCollection.php and EagerField.php to facilitate this new functionality.
Modified MatchFilter.php to accommodate the changes already mentioned. When there's a matchable field, now join statements have been added in the filter() method to make sure relevant records from related entity are included in the search.
Lastly, in EagerField.php, the new method getMatchables() has been introduced to return matchable attributes.