fix: Filtered relations not loaded in relationships and included arrays
A change was made to add functionality to better handle "eager loading" in the MatchFilter class. Prior to this change, the code was using a join operation, which was found to produce empty results in related filter queries. To work around this issue, a subquery was used to circumvent the problematic join method. This ensures that any eager fetches of related data do not result in empty arrays, therefore offering more accurate inclusion and relationship responses.