test: Fix JSONAPI format in tests
This commit addresses refactoring changes to improve the consistency of API response structure.
The change is to put 'attributes' inside 'data' key while sending or receiving data to or from API.
In original structure, data was directly accessed but now it is accessed through 'data.attributes', which is more descriptive and structured.
This approach is aligned with JSON:API specifications and improves readability and understanding of where the actual data lies. So, it is a refactor towards cleaner code and standardized API responses.